% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCM_fit.R
\name{CCM_fit}
\alias{CCM_fit}
\title{Fit a Congruence Class Model (CCM)}
\usage{
CCM_fit(
  Network_stats,
  Prob_Distr,
  Prob_Distr_Params,
  population,
  samplesize = 1000L,
  burnin = 200000L,
  interval = 1000L,
  covPattern = rep(0L, population),
  G = NULL,
  use_G = FALSE,
  partial_network = as.integer(0),
  obs_nodes = NULL,
  Obs_stats = NULL,
  remove_var_last_entry = FALSE
)
}
\arguments{
\item{Network_stats}{Character vector of statistic names.}

\item{Prob_Distr}{Character vector of probability distribution names.}

\item{Prob_Distr_Params}{List of parameter sets for each distribution.}

\item{population}{Integer. Number of nodes.}

\item{samplesize}{Integer. Number of MCMC samples.}

\item{burnin}{Integer. Burn-in period for MCMC.}

\item{interval}{Integer. Thinning interval.}

\item{covPattern}{Integer vector. Covariate pattern or group labels.}

\item{statsonly}{Logical. If TRUE, only return sufficient statistics.}
}
\value{
An object of class \code{CCM_fit} containing:
\itemize{
  \item \code{samples}: MCMC samples of network statistics
  \item \code{theoretical}: theoretical distribution if available
  \item \code{call}: the original function call
}
}
\description{
\code{CCM_fit} fits a Congruence Class Model using MCMC to match
observed network statistics. It takes a list of network statistics and a list
of probability distributions, and returns an object containing MCMC samples,
diagnostics, and theoretical expectations if available.
}
\examples{
CCMnet_python_setup()
population = 100L
fit <- CCM_fit(
  Network_stats = list("Edge"),
  Prob_Distr = list("NP"),
  Prob_Distr_Params = list(dnbinom(0:choose(population,2), size = 1.017340, mu = 6.192894)),
  population = population,
  samplesize = 1000L,
  burnin = 200000L,
  interval = 1000L,
  covPattern = rep(0L, population)  
)

}
