% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCMnet_theoretical_check_degree_mixing.R
\name{CCM_theoretical_check_degree_mixing}
\alias{CCM_theoretical_check_degree_mixing}
\title{Theoretical Degree Distribution + Mixing Check}
\usage{
CCM_theoretical_check_degree_mixing(fit, n_sim)
}
\arguments{
\item{fit}{CCM_fit object}

\item{n_sim}{Number of theoretical samples}
}
\value{
Updated fit object with theoretical degree + mixing distribution
}
\description{
Computes theoretical degree distributions for CCMnet.
}
\keyword{internal}
