% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCMnet_constr_uni_initalstat.R
\name{CCMnet_constr_uni_initalstat}
\alias{CCMnet_constr_uni_initalstat}
\title{Calculate initial network statistics}
\usage{
CCMnet_constr_uni_initalstat(
  Network_stats,
  Prob_Distr,
  Prob_Distr_Params,
  nedges,
  g,
  max_degree,
  population,
  covPattern,
  remove_var_last_entry,
  CCM_constr_info
)
}
\description{
Calculate initial network statistics
}
\keyword{internal}
