% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCMnet_constr_uni_initalstat_mixing_degdist.R
\name{CCMnet_constr_uni_initalstat_mixing_degdist}
\alias{CCMnet_constr_uni_initalstat_mixing_degdist}
\title{Calculate initial statistics}
\usage{
CCMnet_constr_uni_initalstat_mixing_degdist(
  Network_stats,
  Prob_Distr,
  Prob_Distr_Params,
  nedges,
  g,
  max_degree,
  population,
  covPattern,
  remove_var_last_entry,
  CCM_constr_info
)
}
\description{
Calculate initial statistics
}
\keyword{internal}
