% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCMnet_constr_uni_verifyinput_mixing.R
\name{CCMnet_constr_uni_verifyinput_mixing}
\alias{CCMnet_constr_uni_verifyinput_mixing}
\title{Verify input}
\usage{
CCMnet_constr_uni_verifyinput_mixing(
  Network_stats,
  Prob_Distr,
  Prob_Distr_Params,
  nedges,
  g,
  max_degree,
  population,
  covPattern,
  remove_var_last_entry
)
}
\description{
Verify input
}
\keyword{internal}
