% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCMnet_python_setup.R
\name{CCMnet_python_setup}
\alias{CCMnet_python_setup}
\title{Setup CCMnet Python environment}
\usage{
CCMnet_python_setup()
}
\description{
This function configures the Python environment for CCMnet.
}
