% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCM_formula.R
\name{parse_ccm_formulas}
\alias{parse_ccm_formulas}
\title{Parse CCM Model Formulas}
\usage{
parse_ccm_formulas(null_model, alt_model)
}
\arguments{
\item{null_model}{Formula specifying the null CCM.}

\item{alt_model}{Formula specifying alternative terms.}
}
\value{
A list containing the observed network and parsed terms.
}
\description{
Parse CCM Model Formulas
}
\keyword{internal}
