% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.CCM_fit}
\alias{plot.CCM_fit}
\title{Plot CCM_fit object}
\usage{
\method{plot}{CCM_fit}(
  fit,
  stats = NULL,
  type = c("density", "hist"),
  include_theoretical = FALSE,
  ...
)
}
\arguments{
\item{fit}{CCM_fit object}

\item{stats}{Character vector of statistics to plot, e.g., "edges" or c("deg0","deg1")}

\item{type}{"hist" or "density"}

\item{include_theoretical}{Logical. If TRUE, include theoretical distribution in the plot}
}
\value{
ggplot object
}
\description{
Plot CCM_fit object
}
