% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ccm_inference.R
\name{plot.ccm_inference}
\alias{plot.ccm_inference}
\title{Plot CCM Inference Results}
\usage{
\method{plot}{ccm_inference}(x, type = c("test", "network"), term = NULL, ...)
}
\arguments{
\item{x}{A \code{ccm_inference} object.}

\item{type}{Type of plot: \code{"test"} or \code{"network"}.}

\item{term}{Network term to plot.}

\item{...}{Additional arguments (unused).}
}
\description{
Plot CCM Inference Results
}
