% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCMnet_StructAssess_print.ccm_inference.R
\name{print.ccm_inference}
\alias{print.ccm_inference}
\title{Print CCM Inference Object}
\usage{
\method{print}{ccm_inference}(x, ...)
}
\arguments{
\item{x}{A \code{ccm_inference} object.}

\item{...}{Unused.}
}
\description{
Print CCM Inference Object
}
