% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCMnet_tests_degree.R
\name{test_degree_ccm}
\alias{test_degree_ccm}
\title{Degree-Based CCM Test}
\usage{
test_degree_ccm(g_obs, fit_null, population, test_stat)
}
\description{
Degree-Based CCM Test
}
\keyword{internal}
