% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_Prob_Distr_Params.R
\name{Update_Prob_Distr_Params}
\alias{Update_Prob_Distr_Params}
\title{Update Probability Distribution Parameters for CCM MCMC}
\usage{
Update_Prob_Distr_Params(
  g,
  Prob_Distr_Params_hyperprior,
  Network_stats,
  Prob_Distr,
  Prob_Distr_Params,
  G_stats,
  MCMC_wgt,
  population
)
}
\arguments{
\item{g}{An igraph object representing the current network state.}

\item{Prob_Distr_Params_hyperprior}{A list of hyperprior parameters for the probability distributions.}

\item{Network_stats}{A character vector specifying which network statistics to update (e.g., "Degree", "Mixing").}

\item{Prob_Distr}{A list specifying the type of probability distribution used (e.g., "Multinomial_Poisson").}

\item{Prob_Distr_Params}{A list of current probability distribution parameters to be updated.}

\item{G_stats}{A numeric vector of observed network statistics for the current network.}

\item{MCMC_wgt}{Numeric, weight to apply to the network statistics in the update (default = 1).}

\item{population}{Integer, total number of nodes in the network.}
}
\value{
An updated list of probability distribution parameters.
}
\description{
This function updates the probability distribution parameters for network statistics
in a Congruence Class Model (CCM) during MCMC sampling. It supports Bayesian updates
using hyperpriors for degree and mixing distributions.
}
\examples{
\dontrun{
g <- igraph::make_ring(10)
Network_stats <- "Degree"
Prob_Distr <- list("Multinomial_Poisson")
Prob_Distr_Params <- list(1, rep(0.1, 10))
Prob_Distr_Params_hyperprior <- list(
  "Dirichlet_Gamma",
  c(1, 10),
  rep(0.01, 10)
)
G_stats <- tabulate(igraph::degree(g) + 1, nbins = 10)
updated_params <- Update_Prob_Distr_Params(
  g = g,
  Prob_Distr_Params_hyperprior = Prob_Distr_Params_hyperprior,
  Network_stats = Network_stats,
  Prob_Distr = Prob_Distr,
  Prob_Distr_Params = Prob_Distr_Params,
  G_stats = G_stats,
  MCMC_wgt = 1,
  population = 10
)
}

}
