% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCM.R
\name{CCM}
\alias{CCM}
\title{CCM-based Network Inference}
\usage{
CCM(
  null_model,
  alt_model = NULL,
  population,
  Prob_Distr,
  Prob_Distr_Params,
  covPattern,
  samplesize,
  burnin,
  interval,
  test_stat = c("hellinger")
)
}
\arguments{
\item{null_model}{A formula of the form \code{g ~ edges}, where \code{g} is
an \code{igraph} object.}

\item{alt_model}{A one-sided formula specifying network properties to test
(e.g., \code{~ degree}). Default is \code{NULL}.}

\item{population}{Number of nodes in the network.}

\item{Prob_Distr}{Probability distribution used by \code{CCM_fit}.}

\item{Prob_Distr_Params}{Parameters for \code{Prob_Distr}.}

\item{covPattern}{Covariate pattern passed to \code{CCM_fit}.}

\item{samplesize}{Number of MCMC samples.}

\item{burnin}{Number of burn-in iterations.}

\item{interval}{Thinning interval.}

\item{test_stat}{Test statistic to use. Currently only \code{"hellinger"}.}
}
\value{
An object of class \code{"ccm_inference"}.
}
\description{
Fits a null Congruence Class Model (CCM) to an observed network
and conducts simulation-based hypothesis tests for additional network
properties.
}
