% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{ccm_properties}
\alias{ccm_properties}
\title{Network Properties in CCMnet}
\description{
\pkg{CCMnet} implements a systematic hierarchy of network properties 
motivated by the dk-series framework (Mahadevan et al., 2006; Orsini et al., 2015). 
These properties can be targeted using soft constraints governed by 
user-specified probability distributions.
}
\section{Topological-based Properties (dk-series)}{

The following properties follow the dk-series framework, which defines 
increasingly constrained random graph ensembles:
\describe{
  \item{\code{"edges"} (0k)}{The total number of edges within the graph. This 
  is the simplest topological constraint.}
  \item{\code{"degreedist"} (1k)}{A vector where the j-th entry represents 
  the number of nodes having degree j.}
  \item{\code{"degmixing"} (2k)}{The joint degree distribution, represented 
  as a degree mixing matrix. The (i,j) entry represents the number of edges 
  between nodes with degree i and degree j, capturing degree-degree 
  correlations (degree assortativity).}
  \item{\code{"degmixing" + "triangles"} (2.1k)}{Extends the 2k distribution by including 
  the total number of closed triads (triangles) in the network, allowing for 
  clustering constraints.}
}
}

\section{Covariate-based Properties}{

\pkg{CCMnet} incorporates node-level attributes by defining congruence classes 
based on categorical or discretized covariates (e.g., homophily or social 
stratification). 
\describe{
  \item{\code{"mixing"}}{The attribute mixing matrix. The (i,j) entry 
  represents the number of edges between nodes belonging to covariate groups 
  $i$ and $j$. Requires \code{cov_pattern}.}
  \item{\code{"degreedist+degreedist+mixing"}}{A combination property that includes the 
  attribute mixing matrix alongside separate degree distributions calculated 
  for each distinct covariate group. Currently on a binary covariate is implemented. 
  Therefore, input is characteristics for two degree distributions and one scalar mixing value.}
}
}

\references{
Mahadevan, P., Krioukov, D., Fomenkov, K., Huffaker, B., & Vahdat, A. (2006). 
Systematic topology analysis and generation using degree correlations. 
\emph{ACM SIGCOMM Computer Communication Review}.

Orsini, C., et al. (2015). Quantifying randomness in real networks. 
\emph{Nature Communications}.
}
\seealso{
\code{\link{ccm_distributions}}, \code{\link{sample_ccm}}

Other ccm_core: 
\code{\link{ccm_distributions}},
\code{\link{sample_ccm}()}
}
\concept{ccm_core}
