% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCM_theoretical_check.R
\name{CCM_theoretical_check}
\alias{CCM_theoretical_check}
\title{Compare MCMC Estimates with Theoretical Distributions}
\usage{
CCM_theoretical_check(fit, n_sim = nrow(fit$mcmc_stats))
}
\arguments{
\item{fit}{An object returned by \code{CCM_fit}.}

\item{n_sim}{The number of samples drawn from the theoretical distribution}
}
\value{
A diagnostic plot comparing empirical and theoretical distributions.
}
\description{
This function compares the empirical MCMC distribution from \code{CCM_fit}
with the theoretical distribution from the CCM model. It is used as a diagnostic
tool to evaluate model adequacy.
}
\examples{
CCMnet_python_setup()
population = 100L
fit <- CCM_fit(
  Network_stats = list("Edge"),
  Prob_Distr = list("NP"),
  Prob_Distr_Params = list(dnbinom(0:choose(population,2), size = 1.017340, mu = 6.192894)),
  population = population,
  samplesize = 1000L,
  burnin = 200000L,
  interval = 1000L,
  covPattern = rep(0L, population)  
)
CCM_theoretical_check(fit, n_sim = nrow(fit$mcmc_stats))

}
