% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCMnet_theoretical_check_degmix.R
\name{CCM_theoretical_check_degmix}
\alias{CCM_theoretical_check_degmix}
\title{Theoretical Mixing Distribution Check}
\usage{
CCM_theoretical_check_degmix(fit, n_sim)
}
\arguments{
\item{fit}{CCM_fit object}

\item{n_sim}{Number of theoretical samples}
}
\value{
Updated fit object with theoretical mixing distribution
}
\description{
Computes theoretical mixing distributions for CCMnet.
}
\keyword{internal}
