% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCMnet_theoretical_check_degree.R
\name{CCM_theoretical_check_degree}
\alias{CCM_theoretical_check_degree}
\title{Theoretical Degree Distribution Check}
\usage{
CCM_theoretical_check_degree(fit, n_sim)
}
\arguments{
\item{fit}{CCM_fit object}

\item{n_sim}{Number of theoretical samples}
}
\value{
Updated fit object with theoretical degree distribution
}
\description{
Computes theoretical degree distributions for CCMnet.
}
\keyword{internal}
