% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCMnet_diagnostics_traceplot.R
\name{CCM_traceplot}
\alias{CCM_traceplot}
\title{Plot MCMC Trace for CCM Fit}
\usage{
CCM_traceplot(object, stats = NULL, ...)
}
\arguments{
\item{...}{Additional arguments passed to \code{plot()}.}

\item{fit}{A \code{CCM_fit} object.}

\item{stat_index}{Integer. Which statistic to plot.}
}
\value{
A trace plot for the selected MCMC chain.
}
\description{
\code{CCM_traceplot} produces a trace plot of the MCMC samples from
\code{CCM_fit}. This is used to diagnose convergence and mixing.
}
\examples{
CCMnet_python_setup()
population = 100L
fit <- CCM_fit(
  Network_stats = list("Edge"),
  Prob_Distr = list("NP"),
  Prob_Distr_Params = list(dnbinom(0:choose(population,2), size = 1.017340, mu = 6.192894)),
  population = population,
  samplesize = 1000L,
  burnin = 200000L,
  interval = 1000L,
  covPattern = rep(0L, population)  
)
CCM_traceplot(fit, stats = "edges")

}
