% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCM_utils.R
\name{fit_ccm_null}
\alias{fit_ccm_null}
\title{Fit Null CCM}
\usage{
fit_ccm_null(
  null_terms,
  population,
  Prob_Distr,
  Prob_Distr_Params,
  covPattern,
  samplesize,
  burnin,
  interval,
  alt_terms
)
}
\description{
Fit Null CCM
}
\keyword{internal}
