% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCM_utils.R
\name{hellinger_distance}
\alias{hellinger_distance}
\title{Hellinger Distance}
\usage{
hellinger_distance(p, q)
}
\arguments{
\item{p, q}{Probability vectors.}
}
\value{
Hellinger distance between \code{p} and \code{q}.
}
\description{
Hellinger Distance
}
\keyword{internal}
