% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCM_tests_degree.R
\name{run_ccm_tests}
\alias{run_ccm_tests}
\title{Run CCM Hypothesis Tests}
\usage{
run_ccm_tests(g_obs, fit_null, alt_terms, population, test_stat)
}
\description{
Run CCM Hypothesis Tests
}
\keyword{internal}
